+++ src/lib/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./useMediaQuery";
 export * from "./useToast";
 export * from "./useEndpoint";
 export * from "./useUserKey";
+export * from "./useDummyWallet";